+++ corpus/issue-2256/foo.clj
@@ -1,0 +1,7 @@
+(ns foo)
+
+(defmacro bar []
+  )
+
+(bar)
+